Autonomous Commands & Shuffleboard Chooser
