+++ imgui.ini
@@ -40,10 +40,10 @@ Collapsed=0
 
 [MainWindow][GLOBAL]
 width=1280
-height=720
+height=754
 maximized=0
-xpos=459
-ypos=200
+xpos=577
+ypos=233
 userScale=2
 style=0
 
@@ -94,22 +94,22 @@ name=
 [Data Sources][NT:/LiveWindow/Ungrouped/PowerDistributionPanel[0]/Chan5]
 name=
 
-[Data Sources][NT:/LiveWindow/Ungrouped/PowerDistributionPanel[0]/Chan6]
-name=
-
 [Data Sources][NT:/LiveWindow/Ungrouped/PowerDistributionPanel[0]/Chan7]
-name=
-
-[Data Sources][NT:/LiveWindow/.status/LW Enabled]
-name=
-
-[Data Sources][navX-Sensor[0]-Pitch]
 name=
 
 [Data Sources][NT:/LiveWindow/Ungrouped/PowerDistributionPanel[0]/Chan8]
 name=
 
 [Data Sources][NT:/LiveWindow/Ungrouped/PowerDistributionPanel[0]/Chan9]
+name=
+
+[Data Sources][NT:/LiveWindow/.status/LW Enabled]
+name=
+
+[Data Sources][NT:/LiveWindow/DriveTrain/.hasCommand]
+name=
+
+[Data Sources][NT:/LiveWindow/Ungrouped/PowerDistributionPanel[0]/Chan6]
 name=
 
 [Data Sources][NT:/LiveWindow/Ungrouped/PIDController[2]/p]
@@ -121,16 +121,13 @@ name=
 [Data Sources][NT:/FMSInfo/StationNumber]
 name=
 
-[Data Sources][NT:/LiveWindow/DriveTrain/.hasCommand]
-name=
-
 [Data Sources][Y Accel[0]]
 name=
 
 [Data Sources][NT:/LiveWindow/Ungrouped/DifferentialDrive[1]/.actuator]
 name=
 
-[Data Sources][NT:/LiveWindow/Ungrouped/PowerDistributionPanel[0]/Chan4]
+[Data Sources][navX-Sensor[0]-Pitch]
 name=
 
 [Data Sources][FMS:MatchTime]
@@ -145,6 +142,9 @@ name=
 [Data Sources][NT:/SmartDashboard/Robot Speed]
 name=
 
+[Data Sources][NT:/LiveWindow/Ungrouped/PowerDistributionPanel[0]/Chan4]
+name=
+
 [Data Sources][NT:/LiveWindow/Ungrouped/PIDController[2]/i]
 name=
 
@@ -157,13 +157,22 @@ name=
 [Data Sources][NT:/LiveWindow/Ungrouped/PIDController[1]/d]
 name=
 
+[Data Sources][NT:/LiveWindow/Ungrouped/navX-Sensor/.controllable]
+name=
+
 [Data Sources][NT:/LiveWindow/Ungrouped/DifferentialDrive[1]/Right Motor Speed]
 name=
 
 [Data Sources][navX-Sensor[0]-Yaw]
 name=
 
+[Data Sources][NT:/SmartDashboard/Auto Chooser/.instance]
+name=
+
 [Data Sources][NT:/LiveWindow/Ungrouped/PIDController[1]/i]
+name=
+
+[Data Sources][NT:/LiveWindow/Ungrouped/Talon SRX [3]/.controllable]
 name=
 
 [Data Sources][NT:/LiveWindow/Ungrouped/PIDController[2]/setpoint]
@@ -172,7 +181,16 @@ name=
 [Data Sources][NT:/LiveWindow/Ungrouped/PowerDistributionPanel[0]/Voltage]
 name=
 
+[Data Sources][NT:/LiveWindow/Ungrouped/PowerDistributionPanel[0]/.controllable]
+name=
+
 [Data Sources][NT:/LiveWindow/Ungrouped/PIDController[1]/p]
+name=
+
+[Data Sources][NT:/SmartDashboard/Auto Chooser/.controllable]
+name=
+
+[Data Sources][NT:/LiveWindow/DriveTrain/.controllable]
 name=
 
 [Data Sources][Z Accel[0]]
@@ -200,6 +218,9 @@ name=
 name=
 
 [Data Sources][navX-Sensor[0]-Rate]
+name=
+
+[Data Sources][NT:/LiveWindow/Ungrouped/Scheduler/.controllable]
 name=
 
 [Data Sources][FMS:AllianceStationID]
@@ -268,6 +289,9 @@ name=
 [Data Sources][NT:/LiveWindow/Ungrouped/PowerDistributionPanel[0]/TotalCurrent]
 name=
 
+[Data Sources][NT:/LiveWindow/Ungrouped/DifferentialDrive[1]/.controllable]
+name=
+
 [Data Sources][NT:/LiveWindow/Ungrouped/navX-Sensor/Value]
 name=
 
@@ -275,6 +299,9 @@ name=
 name=
 
 [Data Sources][NT:/SmartDashboard/Field/.controllable]
+name=
+
+[Data Sources][NT:/LiveWindow/Ungrouped/Talon SRX [1]/.controllable]
 name=
 
 [Data Sources][NT:/LiveWindow/Ungrouped/DifferentialDrive[1]/Left Motor Speed]
@@ -329,6 +356,9 @@ name=PIDController
 
 [NTProvider][/LiveWindow/Ungrouped/PIDController[2]]
 name=PIDController
+
+[NTProvider][/SmartDashboard/Auto Chooser]
+name=String Chooser
 
 [NTProvider][/LiveWindow/Ungrouped/Scheduler]
 name=Scheduler
