+++ src/main/java/frc/robot/commands/autonomous/drive/DriveCurve.java
@@ -5,34 +5,42 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands.autonomous;
+package frc.robot.commands.autonomous.drive;
 
 import edu.wpi.first.wpilibj.SlewRateLimiter;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.subsystems.DriveTrain;
 
-public class DriveRotate extends CommandBase {
+public class DriveCurve extends CommandBase {
     private final DriveTrain m_drive;
-    private final double degrees, rotationSpeed;
+    private final double m_distance;
+    private final double m_speed;
+    private final double m_rotation;
     private final SlewRateLimiter filter;
-    private double initialRotation;
 
-    public DriveRotate(DriveTrain drive, double degrees, double rotationSpeed) {
-        this.degrees = degrees;
-        this.rotationSpeed = rotationSpeed;
+    /**
+     * Creates a new DriveDistance.
+     *
+     * @param inches The number of inches the robot will drive
+     * @param speed The speed at which the robot will drive
+     * @param drive The drive subsystem on which this command will run
+     */
+    public DriveCurve(DriveTrain drive, double inches, double speed, double rotation) {
+        m_distance = inches;
+        m_speed = speed;
         m_drive = drive;
+        m_rotation = rotation;
         addRequirements(drive);
 
         filter = new SlewRateLimiter(3);
     }
 
-    /**
-     * Called when the command is initially scheduled.
+    /*
+     * Runs when the command is initially scheduled.
      */
     @Override
     public void initialize() {
-        m_drive.resetGyro();
-        initialRotation = m_drive.getAngle().getDegrees();
+        m_drive.resetEncoders();
     }
 
     /**
@@ -40,13 +48,13 @@ public class DriveRotate extends CommandBase {
      */
     @Override
     public void execute() {
-        m_drive.arcadeDrive(0, -filter.calculate(rotationSpeed));
+        m_drive.arcadeDrive(filter.calculate(m_speed), filter.calculate(m_rotation));
     }
 
     /**
      * Called once the command ends or is interrupted.
      *
-     * @param interrupted whether the command was interrupted by another one
+     * @param interrupted If the command was interrupted.
      */
     @Override
     public void end(boolean interrupted) {
@@ -56,11 +64,10 @@ public class DriveRotate extends CommandBase {
     /**
      * Returns true when the command should end.
      *
-     * @return whether the command should end
+     * @return If the command should end.
      */
     @Override
     public boolean isFinished() {
-        return m_drive.getAngle().getDegrees() >= initialRotation + degrees;
-        // return false;
+        return m_drive.getAverageEncoderDistance() >= m_distance;
     }
 }
